% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHWdetection-package.R
\docType{package}
\name{MHWdetection-package}
\alias{MHWdetection}
\alias{MHWdetection-package}
\title{MHWdetection: Detecting marine heatwaves}
\description{
This package serves as a tidy repository for the code used in the paper 'Detecting marine heatwaves'.
}
\author{
\strong{Maintainer}: Robert W. Schlegel \email{robwschlegel@gmail.com} (0000-0002-0705-1287) [contributor]

Authors:
\itemize{
  \item Eric C. J. Oliver (0000-0002-4006-2826)
  \item Alistair J. Hobday
  \item Albertus J. Smit (0000-0002-3799-6126) [contributor]
}

}
\keyword{internal}
