% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder.R
\name{placeholder}
\alias{placeholder}
\title{A placeholder function so pkgdown can compile.}
\usage{
placeholder(x)
}
\arguments{
\item{x}{A single integer or numeric value.}
}
\value{
The function will return double whatever integer is put into it.
}
\description{
This isn't sriously meant to be used for anything, It is just here to trick
pkgdown into functioning more like blgdown.
}
\examples{
res <- placeholder(24)
}
\author{
Robert W. Schlegel
}
